fix: rename invalid :activityId path to support Windows git operations
