--- src/components/CommunityPortal/Activities/:activityid/Resources.css
@@ -1,93 +1,0 @@
-.resources-usage {
-  background: white;
-  padding: 20px;
-  border-radius: 8px;
-  font-family: Arial, sans-serif;
-  width: 100%;
-}
-
-/* border for all the row data */
-.resource-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 10px;
-  border: 2px solid black;
-}
-
-/* Column Styling  for resources*/
-.column {
-  flex: 1;
-  padding: 15px;
-  text-align: center;
-  font-weight: bold;
-}
-
-/* Header Styling */
-.header {
-  background-color: #f5f5f5;
-  font-weight: bold;
-}
-
-.resource-row .column:last-child {
-  border-right: none;
-}
-
-/* colors for text based on status  */
-.status-green {
-  color: green;
-}
-
-.status-yellow {
-  color: orange;
-}
-
-.status-red {
-  color: red;
-}
-
-.view-details {
-  display: block;
-  width: 100%;
-  background-color: #e6f9e6;
-  color: #146c43;
-  font-weight: bold;
-  border-radius: 5px;
-  padding: 10px;
-  cursor: pointer;
-  border: none;
-  text-decoration: none;
-  transition: background 0.3s ease-in-out;
-  text-align: center;
-}
-
-.view-details:hover {
-  background-color: #d1f0d1;
-}
-
-/* Actions Column */
-.action-column {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/* Tick Box */
-.tick-box {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 20px;
-}
-
-.tick-label {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: bold;
-  background-color: white;
-  border: 2px solid black;
-  padding: 10px 15px;
-  border-radius: 8px;
-  cursor: pointer;
-}
